# src/main/resources/api.yaml
openapi: 3.1.0
info:
  title: Mosify Reward System API
  version: 1.0.0
servers:
  - url: '/'
paths:
  /boards:
    get:
      summary: Get all boards
      operationId: getAllBoards
      responses:
        '200':
          description: List of boards
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BoardResponse'
    post:
      summary: Create a board
      operationId: createBoard
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/BoardRequest'
      responses:
        '201':
          description: Board created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BoardResponse'
  /boards/{id}:
    delete:
      summary: Delete board by ID
      operationId: deleteBoard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Board deleted successfully
        '404':
          description: Board not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /boards/{id}/users:
    get:
      summary: Get users belonging to a board
      operationId: getBoardUsers
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of board users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/BoardUserResponse'
  /boards/{id}/users/{userId}:
    post:
      summary: Add an existing user to a board
      operationId: addUserToBoard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: alias
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: User added to board successfully
        '404':
          description: Board or User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Remove a user from a board
      operationId: removeUserFromBoard
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: userId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: User removed from board successfully
        '404':
          description: Board-User membership not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /boards/{id}/categories:
    get:
      summary: Get categories belonging to a board
      operationId: getBoardCategories
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of board categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponse'
  /boards/{id}/tasks:
    get:
      summary: Get tasks belonging to a board
      operationId: getBoardTasks
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of board tasks
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TaskResponse'
  /boards/{id}/transactions:
    get:
      summary: Get transactions belonging to a board
      operationId: getBoardTransactions
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of board transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionResponse'
...
  /categories:
    get:
      summary: Get all categories
      operationId: getAllCategories
      responses:
        '200':
          description: List of categories
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/CategoryResponse'
    post:
      summary: Create a category
      operationId: createCategory
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CategoryRequest'
      responses:
        '201':
          description: Category created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
  /categories/{id}:
    get:
      summary: Get category by ID
      operationId: getCategoryById
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Category details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CategoryResponse'
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
    delete:
      summary: Delete category by ID
      operationId: deleteCategory
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Category deleted successfully
        '404':
          description: Category not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  schemas:
...
    CategoryRequest:
      type: object
      required:
        - userId
        - boardId
        - name
      properties:
        userId:
          type: string
          format: uuid
        boardId:
          type: string
          format: uuid
        name:
          type: string
          example: "Studies"
        description:
          type: string
          example: "Academic tasks and goals"
    CategoryResponse:
      type: object
      required:
        - id
        - userId
        - boardId
        - name
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        boardId:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
...
    TaskResponse:
      type: object
      required:
        - id
        - title
        - categoryId
        - type
        - pointsValue
        - active
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        categoryId:
          type: string
          format: uuid
        type:
          type: string
          enum: [SINGLE_USE, RECURRENT]
        frequency:
          type: string
          enum: [DAILY, WEEKLY, MONTHLY]
          nullable: true
        pointsValue:
          type: integer
        active:
          type: boolean
    TransactionResponse:
      type: object
      required:
        - id
        - userId
        - pointsAffected
        - createdAt
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        taskId:
          type: string
          format: uuid
          nullable: true
        pointsAffected:
          type: integer
        createdAt:
          type: string
          format: date-time
    ErrorResponse:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: "INSUFFICIENT_BALANCE"
        message:
          type: string
          example: "The user has insufficient points balance."
    BoardRequest:
      type: object
      required:
        - name
      properties:
        name:
          type: string
          example: "Family Board"
    BoardResponse:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
    BoardUserResponse:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        pointsBalance:
          type: integer
        alias:
          type: string
        userName:
          type: string
        boardId:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
...